+++ pymongo/uri_parser.py
@@ -23,14 +23,15 @@ try:
 except ImportError:
     _HAVE_DNSPYTHON = False
 
-from bson.py3compat import PY3, string_type
+from bson.py3compat import abc, iteritems, string_type, PY3
 
 if PY3:
     from urllib.parse import unquote_plus
 else:
     from urllib import unquote_plus
 
-from pymongo.common import get_validated_options
+from pymongo.common import (
+    get_validated_options, URI_OPTIONS_DEPRECATION_MAP, INTERNAL_URI_OPTION_NAME_MAP)
 from pymongo.errors import ConfigurationError, InvalidURI
 
 
@@ -39,6 +40,80 @@ SCHEME_LEN = len(SCHEME)
 SRV_SCHEME = 'mongodb+srv://'
 SRV_SCHEME_LEN = len(SRV_SCHEME)
 DEFAULT_PORT = 27017
+
+
+class _CaseInsensitiveDictionary(abc.MutableMapping):
+    def __init__(self, *args, **kwargs):
+        self.__casedkeys = {}
+        self.__data = {}
+        self.update(dict(*args, **kwargs))
+
+    def __contains__(self, key):
+        return key.lower() in self.__data
+
+    def __len__(self):
+        return len(self.__data)
+
+    def __iter__(self):
+        return (self.__casedkeys[key] for key in self.__casedkeys)
+
+    def __repr__(self):
+        return str(self.__data)
+
+    def __setitem__(self, key, value):
+        lc_key = key.lower()
+        self.__casedkeys[lc_key] = key
+        self.__data[lc_key] = value
+
+    def __getitem__(self, key):
+        return self.__data[key.lower()]
+
+    def __delitem__(self, key):
+        lc_key = key.lower()
+        del self.__casedkeys[lc_key]
+        del self.__data[lc_key]
+
+    def get(self, key, default=None):
+        lc_key = key.lower()
+        if lc_key in self:
+            return self.__data[lc_key]
+        return default
+
+    def pop(self, key, *args, **kwargs):
+        lc_key = key.lower()
+        self.__casedkeys.pop(lc_key, None)
+        return self.__data.pop(lc_key, *args, **kwargs)
+
+    def popitem(self):
+        lc_key, cased_key = self.__casedkeys.popitem()
+        value = self.__data.pop(lc_key)
+        return cased_key, value
+
+    def clear(self):
+        self.__casedkeys.clear()
+        self.__data.clear()
+
+    def setdefault(self, key, default=None):
+        lc_key = key.lower()
+        if key in self:
+            return self.__data[lc_key]
+        else:
+            self.__casedkeys[lc_key] = key
+            self.__data[lc_key] = default
+            return default
+
+    def update(self, other):
+        for key in other:
+            self[key] = other[key]
+
+    def cased_key(self, key):
+        return self.__casedkeys[key.lower()]
+
+    def as_dict(self):
+        lc_data = {}
+        for lc_key in self.__data:
+            lc_data[lc_key] = self.__data[lc_key]
+        return lc_data
 
 
 def parse_userinfo(userinfo):
@@ -129,6 +204,66 @@ def parse_host(entity, default_port=DEFAULT_PORT):
     return host.lower(), port
 
 
+_IMPLICIT_TLSINSECURE_OPTS = {"tlsallowinvalidcertificates",
+                              "tlsallowinvalidhostnames"}
+
+
+def _parse_options(opts, delim):
+    """Helper method for split_options which creates the options dict.
+    Also handles the creation of a list for the URI tag_sets/
+    readpreferencetags portion and the use of the tlsInsecure option."""
+    options = _CaseInsensitiveDictionary()
+    for uriopt in opts.split(delim):
+        key, value = uriopt.split("=")
+        if key.lower() == 'readpreferencetags':
+            options.setdefault(key, []).append(value)
+        else:
+            if key in options:
+                warnings.warn("Duplicate URI option '%s'." % (key,))
+            options[key] = unquote_plus(value)
+
+    if 'tlsInsecure' in options:
+        for implicit_option in _IMPLICIT_TLSINSECURE_OPTS:
+            if implicit_option in options:
+                warn_msg = "URI option '%s' overrides value implied by '%s'."
+                warnings.warn(warn_msg % (options.cased_key(implicit_option),
+                                          options.cased_key('tlsInsecure')))
+                continue
+            options[implicit_option] = options['tlsInsecure']
+
+    return options
+
+
+def _handle_option_deprecations(options):
+    """Issue appropriate warnings when deprecated options are present in the
+    options dictionary. Removes deprecated option key, value pairs if the
+    options dictionary is found to also have the renamed option."""
+    undeprecated_options = _CaseInsensitiveDictionary()
+    for key, value in iteritems(options):
+        optname = str(key).lower()
+        if optname in URI_OPTIONS_DEPRECATION_MAP:
+            renamed_key = URI_OPTIONS_DEPRECATION_MAP[optname]
+            if renamed_key.lower() in options:
+                warnings.warn("Deprecated option '%s' ignored in favor of "
+                              "'%s'." % (str(key), renamed_key))
+                continue
+            warnings.warn("Option '%s' is deprecated, use '%s' instead." % (
+                          str(key), renamed_key))
+        undeprecated_options[str(key)] = value
+    return undeprecated_options
+
+
+def _normalize_options(options):
+    """Renames keys in the options dictionary to their internally-used
+    names."""
+    normalized_options = {}
+    for key, value in iteritems(options):
+        optname = str(key).lower()
+        intname = INTERNAL_URI_OPTION_NAME_MAP.get(optname, key)
+        normalized_options[intname] = options[key]
+    return normalized_options
+
+
 def validate_options(opts, warn=False):
     """Validates and normalizes options passed in a MongoDB URI.
 
@@ -138,41 +273,14 @@ def validate_options(opts, warn=False):
 
     :Parameters:
         - `opts`: A dict of MongoDB URI options.
-        - `warn` (optional): If ``True`` then warnigns will be logged and
+        - `warn` (optional): If ``True`` then warnings will be logged and
           invalid options will be ignored. Otherwise invalid options will
           cause errors.
     """
     return get_validated_options(opts, warn)
 
 
-def _parse_options(opts, delim):
-    """Helper method for split_options which creates the options dict.
-    Also handles the creation of a list for the URI tag_sets/
-    readpreferencetags portion."""
-    options = {}
-    for opt in opts.split(delim):
-        key, val = opt.split("=")
-        if key.lower() == 'readpreferencetags':
-            options.setdefault('readpreferencetags', []).append(val)
-        else:
-            # str(option) to ensure that a unicode URI results in plain 'str'
-            # option names. 'normalized' is then suitable to be passed as
-            # kwargs in all Python versions.
-            if str(key) in options:
-                warnings.warn("Duplicate URI option %s" % (str(key),))
-            options[str(key)] = unquote_plus(val)
-
-    # Special case for deprecated options
-    if "wtimeout" in options:
-        if "wtimeoutMS" in options:
-            options.pop("wtimeout")
-        warnings.warn("Option wtimeout is deprecated, use 'wtimeoutMS'"
-                      " instead")
-
-    return options
-
-
-def split_options(opts, validate=True, warn=False):
+def split_options(opts, validate=True, warn=False, normalize=True):
     """Takes the options portion of a MongoDB URI, validates each option
     and returns the options in a dictionary.
 
@@ -180,6 +288,10 @@ def split_options(opts, validate=True, warn=False):
         - `opt`: A string representing MongoDB URI options.
         - `validate`: If ``True`` (the default), validate and normalize all
           options.
+        - `warn`: If ``False`` (the default), suppress all warnings raised
+          during validation of options.
+        - `normalize`: If ``True`` (the default), renames all options to their
+          internally-used names.
     """
     and_idx = opts.find("&")
     semi_idx = opts.find(";")
@@ -197,8 +309,14 @@ def split_options(opts, validate=True, warn=False):
     except ValueError:
         raise InvalidURI("MongoDB URI options are key=value pairs.")
 
+    options = _handle_option_deprecations(options)
+
     if validate:
-        return validate_options(options, warn)
+        options = validate_options(options, warn)
+
+    if normalize:
+        options = _normalize_options(options)
+
     return options